+++ webui/src/components/ConfigurationList.vue
@@ -233,7 +233,6 @@ const saveToServer = () => {
   try {
     for (let property: string in props.configurationValue) {
       if (props.configurationGroups[0].properties[property].form_type == 'password') {
-        console.log(props.configurationGroups[0].properties[property].password)
         props.configurationValue[property] = createHash(
           props.configurationValue[property],
           props.configurationGroups[0].properties[property].password